Intervals refactoring: ListView instead of LinearLayout with a list of views.
